Make the test suite trustable and more robust by replacing the assertion library
The problem was that in Bash the tests didn't fail when the actual value
was empty.

Improvements:
It fails after the first failed assertion.
It fails in bash even if the actual value is empty.
It prints the path with a line number of the failed assertion.

Lost features:
A bunch, but trust is more important than features.